Renamed project and rm mutex in read
