+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2024 SONIA AUV
+Copyright (c) 2025 SONIA AUV
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
