+++ include/rs485_port_manager/RS485Interface.hpp
@@ -22,7 +22,7 @@
 #include "sonia_common_ros2/srv/dropper_service.hpp"
 
 
-namespace sonia_hw_interface
+namespace rs485_port_manager
 {
     /**
      * @brief Internal Queue Object
