+++ include/rs485_port_manager/SharedQueue.hpp
