+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>hardware_interface_manager</name>
+  <name>rs485_port_manager</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="log.club.sonia@etsmtl.net">Sonia LOG Team</maintainer>
