+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "hardware_interface_manager/RS485Interface.hpp"
+#include "rs485_port_manager/RS485Interface.hpp"
 // #include "hardware_interface_manager/ThrusterProvider.hpp"
 #include <stdlib.h>
 
@@ -8,7 +8,7 @@
 int main(int argc, char *argv[])
 {
     rclcpp::init(argc, argv);
-    auto thrust = std::make_shared<sonia_hw_interface::RS485Interface>();
+    auto thrust = std::make_shared<rs485_port_manager::RS485Interface>();
     if (!thrust->OpenPort())
     {
         printf("Could not open port...\n");
